Final Push For The Mouse Accuracy Game
This is the final push for the mouse accuracy game. Unfortunately I did not manage to add a database conenction and the necessary dataaccess functions for the leaderboard to be accurate and persinstent(as it stands it is totally static). Other than that the game, the game settings and the game results work as intended.